+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Unreleased
 
 * Add `buffer` function
-* Add `Signal#buffer` function
+* Add `Signal#buffer` method
+* Add `apply` function
 
 ## 2.0.0 (2019-01-12)